+++ odbc-api/tests/common.rs
@@ -36,7 +36,7 @@ pub fn setup_empty_table(
 ) -> Result<(), odbc_api::Error> {
     let drop_table = &format!("DROP TABLE IF EXISTS {}", table_name);
 
-    let column_names = &["a", "b", "c", "d", "e", "f", "g", "h"];
+    let column_names = &["a", "b", "c", "d", "e", "f", "g", "h", "i"];
     let cols = column_types
         .iter()
         .zip(column_names)